Remove now uneccessary test classes
