+++ game-core/src/main/java/tools/map/making/MapPropertyWrapper.java
@@ -1,5 +1,7 @@
 package tools.map.making;
 
+import static com.google.common.base.Preconditions.checkNotNull;
+
 import java.awt.Color;
 import java.beans.Introspector;
 import java.io.File;
@@ -28,7 +30,6 @@ import games.strategy.engine.data.properties.MapProperty;
 import games.strategy.engine.data.properties.NumberProperty;
 import games.strategy.engine.data.properties.PropertiesUi;
 import games.strategy.engine.data.properties.StringProperty;
-import games.strategy.util.PropertyUtil;
 import games.strategy.util.Tuple;
 import tools.util.ToolLogger;
 
@@ -130,7 +131,7 @@ public class MapPropertyWrapper<T> extends AEditableProperty {
           setter.getName().substring(Math.min(3, setter.getName().length()), setter.getName().length());
 
       final String fieldName = Introspector.decapitalize(propertyName);
-      final Field field = PropertyUtil.getPropertyField(fieldName, object.getClass());
+      final Field field = getPropertyField(fieldName, object.getClass());
       final Object currentValue;
       try {
         currentValue = field.get(object);
@@ -171,6 +172,48 @@ public class MapPropertyWrapper<T> extends AEditableProperty {
   @Override
   public boolean validate(final Object value) {
     return property.validate(value);
+  }
+
+
+  private static Field getFieldIncludingFromSuperClasses(final Class<?> c, final String name,
+      final boolean justFromSuper) {
+    if (!justFromSuper) {
+      try {
+        return c.getDeclaredField(name); // TODO: unchecked reflection
+      } catch (final NoSuchFieldException e) {
+        return getFieldIncludingFromSuperClasses(c, name, true);
+      }
+    }
+
+    if (c.getSuperclass() == null) {
+      throw new IllegalStateException("No such Property Field: " + name);
+    }
+    try {
+      return c.getSuperclass().getDeclaredField(name); // TODO: unchecked reflection
+    } catch (final NoSuchFieldException e) {
+      return getFieldIncludingFromSuperClasses(c.getSuperclass(), name, true);
+    }
+  }
+
+  /**
+   * Gets the backing field for the property with the specified name in the specified type.
+   *
+   * @param propertyName The property name.
+   * @param type The type that hosts the property.
+   *
+   * @return The backing field for the specified property.
+   *
+   * @throws IllegalStateException If no backing field for the specified property exists.
+   */
+  public static Field getPropertyField(final String propertyName, final Class<?> type) {
+    checkNotNull(propertyName);
+    checkNotNull(type);
+
+    try {
+      return getFieldIncludingFromSuperClasses(type, "m_" + propertyName, false);
+    } catch (final IllegalStateException ignored) {
+      return getFieldIncludingFromSuperClasses(type, propertyName, false);
+    }
   }
 
   public static void main(final String[] args) {
